Consistent handling of manifest.mf and feature.xml files
